chore: refactor publisher service
