+++ Cargo.toml
@@ -10,10 +10,6 @@ default-run = "server"
 name = "server"
 path = "src/bin/server.rs"
 
-[[bin]]
-name = "publisher_service"
-path = "src/bin/publisher.rs"
-
 [dependencies]
 wc = { git = "https://github.com/WalletConnect/utils-rs.git", tag = "v0.5.1", features = ["full"] }
 cerberus = { git = "https://github.com/WalletConnect/cerberus.git", tag = "v0.9.1" }
