+++ src/services/mod.rs
@@ -1,5 +1,5 @@
-pub mod private_http;
-pub mod public_http;
+pub mod private_http_server;
+pub mod public_http_server;
 pub mod publisher_service;
 pub mod watcher_expiration_job;
-pub mod websocket_service;
+pub mod websocket_server;
