+++ src/services/private_http_server/handlers/metrics.rs
