+++ src/services/public_http_server/handlers/get_subscribers_v1.rs
