+++ src/services/public_http_server/handlers/health.rs
