+++ src/services/public_http_server/handlers/subscribe_topic.rs
